# WIP SHOWCASE [performance] Tree batch expand