--- a/bundles/org.eclipse.jface/src/org/eclipse/jface/viewers/AbstractTreeViewer.java
+++ b/bundles/org.eclipse.jface/src/org/eclipse/jface/viewers/AbstractTreeViewer.java
@@ -46,6 +46,7 @@ import org.eclipse.swt.events.TreeListener;
 import org.eclipse.swt.graphics.Point;
 import org.eclipse.swt.widgets.Control;
 import org.eclipse.swt.widgets.Item;
+import org.eclipse.swt.widgets.Tree;
 import org.eclipse.swt.widgets.TreeItem;
 import org.eclipse.swt.widgets.Widget;
 
@@ -362,9 +363,7 @@ public abstract class AbstractTreeViewer extends ColumnViewer {
 
 		// Optimize for the empty case
 		if (items.length == 0) {
-			for (Object element : elements) {
-				createTreeItem(widget, element, -1);
-			}
+			createItems(widget, elements);
 			return;
 		}
 
@@ -447,6 +446,13 @@ public abstract class AbstractTreeViewer extends ColumnViewer {
 				}
 			}
 		}
+	}
+
+	/**
+	 * @since 3.33
+	 */
+	protected void createItems(Widget widget, Object[] elements) {
+		// to be overriden!
 	}
 
 	/**
@@ -846,9 +852,7 @@ public abstract class AbstractTreeViewer extends ColumnViewer {
 				} else {
 					children = getSortedChildren(parentElement);
 				}
-				for (Object element : children) {
-					createTreeItem(widget, element, -1);
-				}
+				createItems(widget, children);
 			}
 		} finally {
 			setBusy(oldBusy);
@@ -856,16 +860,16 @@ public abstract class AbstractTreeViewer extends ColumnViewer {
 	}
 
 	/**
-	 * Creates a single item for the given parent and synchronizes it with the
-	 * given element.
+	 * Creates a single item for the given parent and synchronizes it with the given
+	 * element. The fastest way to insert many items is documented in
+	 * {@link TreeItem#TreeItem(Tree,int,int)}
 	 *
-	 * @param parent
-	 *            the parent widget
-	 * @param element
-	 *            the element
-	 * @param index
-	 *            if non-negative, indicates the position to insert the item
-	 *            into its parent
+	 * @param parent  the parent widget
+	 * @param element the element
+	 * @param index   if non-negative, indicates the position to insert the item
+	 *                into its parent
+	 * @see org.eclipse.swt.widgets.TreeItem#TreeItem(org.eclipse.swt.widgets.Tree,
+	 *      int, int)
 	 */
 	protected void createTreeItem(Widget parent, Object element, int index) {
 		Item item = newItem(parent, SWT.NULL, index);
@@ -1831,6 +1835,7 @@ public abstract class AbstractTreeViewer extends ColumnViewer {
 				return;
 			}
 			createChildren(widget, false);
+			// XXX for performance widget should be expanded after expanding children:
 			if (widget instanceof Item) {
 				setExpanded((Item) widget, true);
 			}
@@ -1844,6 +1849,7 @@ public abstract class AbstractTreeViewer extends ColumnViewer {
 					}
 				}
 			}
+			// XXX expanding here fails on linux
 		}
 	}