[win]:
OLD implementation:
3 ms to process 10000 items
5 ms to process 100000 items
The time to select scales as O(n^0,19) as the size of the model grows from 10000 to 100000
1068 ms to process 10000 items
10672 ms to process 100000 items
The time to reveal scales as O(n^1,00) as the size of the model grows from 10000 to 100000

NEW batch implementation:
3 ms to process 10000 items
20 ms to process 100000 items  <----- WORSE
The time to select scales as O(n^0,75) as the size of the model grows from 10000 to 100000
1053 ms to process 10000 items
10487 ms to process 100000 items
The time to reveal scales as O(n^1,00) as the size of the model grows from 10000 to 100000