@jukzi 2023-12-01
> Well, if the experimant is a failure, lets reject it.

i wonder if you want to test it on MAC/linux