@basilevs 2023-12-01
No, if this is a regression on Windows, the only way this would be useful, is to add platform checks in TreeViewer, which would be unacceptable.